Fix auto-hiding nav in mobile safari
The nav is designed to switch to a visible, fixed state when scrolling
up the page on mobile. In mobile safari, however, scrolling up via
intertial scrolling will trigger the toolbars to display, which changes
the scrollTop slightly, making it appear that the user is scrolling down
momentarily when they are actually scrolling up. This caused the nav to
incorrectly switch to a hidden state. Checking for window.innerHeight
consistency prevents that.

Only works about 90% of the time but it's better than nothing.